Render Attachments in Threads
